fix middleware interface naming issues in tests
