+++ tests/MiddlewareRespondStub.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Idealo\Middleware\Tests;
+
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+
+class MiddlewareResponseStub implements MiddlewareInterface
+{
+    /**
+     * @var ResponseInterface
+     */
+    private $response;
+
+    public function __construct(ResponseInterface $response)
+    {
+        $this->response = $response;
+    }
+
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+    {
+        $request->getBody();
+
+        return $this->response;
+    }
+}
