+++ tests/MiddlewareStub.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Idealo\Middleware\Tests;
+
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+
+class MiddlewareStub implements MiddlewareInterface
+{
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+    {
+        $request->getBody();
+
+        return $handler->handle($request);
+    }
+}
